requirements.txt
aiohttp
numpy
openai
python_dotenv
requests
scipy
tiktoken
torch
transformers
